マウスの入力状態を更新する関数を追加

+++ Library/DxLibHelper/Vector/vector2_int.hpp
@@ -24,7 +24,7 @@ struct Vector2Int
 	/// @return 大きさ
 	[[nodiscard]] float GetMagnitude() const
 	{
-		return sqrt(x * x + y * y);
+		return sqrtf(static_cast<float>(x * x + y * y));
 	}
 
 	/// @brief 大きさの二乗を取得する
